api/requirements.txt
Django==3.1
djangorestframework==3.11.1
psycopg2==2.8.5
twilio==6.44.1
django-cors-headers==3.4.0